requirements.txt
torch
torchvision
tqdm
matplotlib
numpy
scipy
pillow
onnx
opencv-python
pycaret
streamlit